Тест. Попытка сделать приложение основным для SMS.
